NF: First try of Fast-LTS implementation (large sample LTS).
The fit_large_sample method has been created. It has yet to be modified
in order to take into account a complete subsampling scheme (thus, it would
become a correct Fast-LTS implementation).
The random_fit method has been modified so that we can stop after stage 1 and
hence release the best candidates obtained so far without finishing the run.
The same method can now start from an initial set of subsets, given by their
masks (instead of randomly drawing the subsets). The way the candidate are
stored has been modified so it is easier to pool the results from the
different subsets to a larger set.

A benchmark script has been added in examples. Unfortunately, I was not able
to show a real benefit of using the large sample LTS version. It is even worse
sometime.
Parameters of the simple LTS has been adjusted so that they correspond roughly
to what is done with in the alrge sample version (n_keep and max_starts
parameters have been set to the same values in the example).